requirements.txt
asyncpg==0.30.0
psycopg2-binary==2.9.10
requests==2.31.0
beautifulsoup4==4.12.3
Pillow==10.4.0
aiohttp==3.11.0
python-dotenv==1.0.1
ddgs
python-telegram-bot[job-queue]==20.7
httpx==0.25.2
anyio==3.7.1